Moved BabelFishAction definition to separate header in preparation for BabelFishActionServer.
